creates hook for handling keypresses
