+++ src/renderer/src/components/layout/Layout.tsx
@@ -3,8 +3,10 @@ import { ImagePackage } from 'main/types';
 
 import { useApp } from '../context/app-context';
 import SelectFolder from '../select-folder/SelectFolder';
-import './Layout.css';
 import BigPreview from '../big-preview/BigPreview';
+
+import './Layout.css';
+import useKeyHandlers from './useKeyHandlers';
 
 const Layout = () => {
   const {
@@ -20,34 +22,7 @@ const Layout = () => {
     ensureFolderPath();
   }, []);
 
-  useEffect(() => {
-    const handleKeyPress = (e: KeyboardEvent) => {
-      e.preventDefault();
-
-      const index = images.findIndex((image) => image.id === selectedImage?.id);
-
-      if (index === -1) return;
-      switch (e.key) {
-        case 'ArrowLeft':
-          if (index === 0) return;
-          setSelectedImage(images[index - 1]);
-          break;
-        case 'ArrowRight':
-          if (index === images.length - 1) return;
-          setSelectedImage(images[index + 1]);
-          break;
-        default:
-          break;
-      }
-    };
-
-    window.addEventListener('keydown', handleKeyPress);
-
-    // Remove the event listener when the component unmounts
-    return () => {
-      window.removeEventListener('keydown', handleKeyPress);
-    };
-  }, [selectedImage, images]);
+  useKeyHandlers();
 
   const handleImageClick = (imagePackage: ImagePackage) => {
     setSelectedImage(imagePackage);
